# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FunScript Hub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Creator {
  id: number;
  name: string;
  handle: string;
  /** @nullable */
  bio?: string | null;
  scriptCount?: number;
  totalViews?: number;
  createdAt: string;
  updatedAt?: string;
}

export interface CreatorInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  handle: string;
  bio?: string;
}

export interface CreatorUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  handle?: string;
  /** @nullable */
  bio?: string | null;
}

export interface FunScript {
  id: number;
  creatorId: number;
  /** @nullable */
  creatorName?: string | null;
  /** @nullable */
  creatorHandle?: string | null;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  contentUrl?: string | null;
  shareToken: string;
  viewCount: number;
  /** @nullable */
  expiresAt?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface FunScriptInput {
  creatorId: number;
  /** @minLength 1 */
  title: string;
  description?: string;
  contentUrl?: string;
  expiresAt?: string;
}

export interface FunScriptUpdate {
  /** @minLength 1 */
  title?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  contentUrl?: string | null;
  /** @nullable */
  expiresAt?: string | null;
}

export interface AccessLog {
  id: number;
  funScriptId: number;
  /** @nullable */
  ipAddress?: string | null;
  /** @nullable */
  userAgent?: string | null;
  /** @nullable */
  referrer?: string | null;
  accessedAt: string;
}

export interface AccessLogWithScript {
  id: number;
  funScriptId: number;
  /** @nullable */
  scriptTitle?: string | null;
  /** @nullable */
  shareToken?: string | null;
  /** @nullable */
  ipAddress?: string | null;
  /** @nullable */
  userAgent?: string | null;
  /** @nullable */
  referrer?: string | null;
  accessedAt: string;
}

export interface DailyCount {
  date: string;
  count: number;
}

export interface ScriptStats {
  totalViews: number;
  uniqueIps: number;
  dailyCounts: DailyCount[];
}

export interface OverviewStats {
  totalCreators: number;
  totalScripts: number;
  totalViews: number;
  viewsToday: number;
}