# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FunScript Hub API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: creators
    description: Creator operations
  - name: scripts
    description: FunScript operations
  - name: access
    description: Public access tracking
  - name: stats
    description: Stats and analytics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /creators:
    get:
      operationId: listCreators
      tags: [creators]
      summary: List all creators
      responses:
        "200":
          description: List of creators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Creator"
    post:
      operationId: createCreator
      tags: [creators]
      summary: Create a new creator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatorInput"
      responses:
        "201":
          description: Creator created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Creator"
        "400":
          description: Invalid input

  /creators/{id}:
    get:
      operationId: getCreator
      tags: [creators]
      summary: Get a creator by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Creator found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Creator"
        "404":
          description: Creator not found
    patch:
      operationId: updateCreator
      tags: [creators]
      summary: Update a creator
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatorUpdate"
      responses:
        "200":
          description: Creator updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Creator"
        "404":
          description: Creator not found
    delete:
      operationId: deleteCreator
      tags: [creators]
      summary: Delete a creator
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Creator deleted
        "404":
          description: Creator not found

  /creators/{creatorId}/scripts:
    get:
      operationId: listScriptsByCreator
      tags: [scripts]
      summary: List FunScripts for a creator
      parameters:
        - name: creatorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of FunScripts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FunScript"

  /scripts:
    get:
      operationId: listScripts
      tags: [scripts]
      summary: List all FunScripts
      responses:
        "200":
          description: List of all FunScripts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FunScript"
    post:
      operationId: createScript
      tags: [scripts]
      summary: Create a new FunScript
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FunScriptInput"
      responses:
        "201":
          description: FunScript created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FunScript"
        "400":
          description: Invalid input

  /scripts/{id}:
    get:
      operationId: getScript
      tags: [scripts]
      summary: Get a FunScript by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: FunScript found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FunScript"
        "404":
          description: FunScript not found
    patch:
      operationId: updateScript
      tags: [scripts]
      summary: Update a FunScript
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FunScriptUpdate"
      responses:
        "200":
          description: FunScript updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FunScript"
        "404":
          description: FunScript not found
    delete:
      operationId: deleteScript
      tags: [scripts]
      summary: Delete a FunScript
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: FunScript deleted
        "404":
          description: FunScript not found

  /scripts/{id}/accesses:
    get:
      operationId: listScriptAccesses
      tags: [access]
      summary: List access logs for a FunScript
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Access logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccessLog"

  /scripts/{id}/stats:
    get:
      operationId: getScriptStats
      tags: [stats]
      summary: Get access stats for a FunScript (daily counts for last 30 days)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stats for the FunScript
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScriptStats"

  /s/{token}:
    get:
      operationId: accessScript
      tags: [access]
      summary: Access a FunScript by share token (tracks the visit and returns script info)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: FunScript accessed and tracked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FunScript"
        "404":
          description: FunScript not found

  /stats/overview:
    get:
      operationId: getOverviewStats
      tags: [stats]
      summary: Get dashboard overview stats
      responses:
        "200":
          description: Overview stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverviewStats"

  /stats/top-scripts:
    get:
      operationId: getTopScripts
      tags: [stats]
      summary: Get top FunScripts by view count
      responses:
        "200":
          description: Top scripts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FunScript"

  /stats/recent-accesses:
    get:
      operationId: getRecentAccesses
      tags: [stats]
      summary: Get most recent access log entries across all scripts
      responses:
        "200":
          description: Recent access entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccessLogWithScript"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Creator:
      type: object
      required: [id, name, handle, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        handle:
          type: string
        bio:
          type: ["string", "null"]
        scriptCount:
          type: integer
        totalViews:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string

    CreatorInput:
      type: object
      required: [name, handle]
      properties:
        name:
          type: string
          minLength: 1
        handle:
          type: string
          minLength: 1
        bio:
          type: string

    CreatorUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        handle:
          type: string
          minLength: 1
        bio:
          type: ["string", "null"]

    FunScript:
      type: object
      required: [id, creatorId, title, shareToken, viewCount, createdAt]
      properties:
        id:
          type: integer
        creatorId:
          type: integer
        creatorName:
          type: ["string", "null"]
        creatorHandle:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        contentUrl:
          type: ["string", "null"]
        shareToken:
          type: string
        viewCount:
          type: integer
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    FunScriptInput:
      type: object
      required: [creatorId, title]
      properties:
        creatorId:
          type: integer
        title:
          type: string
          minLength: 1
        description:
          type: string
        contentUrl:
          type: string
        expiresAt:
          type: string

    FunScriptUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        contentUrl:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]

    AccessLog:
      type: object
      required: [id, funScriptId, accessedAt]
      properties:
        id:
          type: integer
        funScriptId:
          type: integer
        ipAddress:
          type: ["string", "null"]
        userAgent:
          type: ["string", "null"]
        referrer:
          type: ["string", "null"]
        accessedAt:
          type: string

    AccessLogWithScript:
      type: object
      required: [id, funScriptId, accessedAt]
      properties:
        id:
          type: integer
        funScriptId:
          type: integer
        scriptTitle:
          type: ["string", "null"]
        shareToken:
          type: ["string", "null"]
        ipAddress:
          type: ["string", "null"]
        userAgent:
          type: ["string", "null"]
        referrer:
          type: ["string", "null"]
        accessedAt:
          type: string

    DailyCount:
      type: object
      required: [date, count]
      properties:
        date:
          type: string
        count:
          type: integer

    ScriptStats:
      type: object
      required: [totalViews, uniqueIps, dailyCounts]
      properties:
        totalViews:
          type: integer
        uniqueIps:
          type: integer
        dailyCounts:
          type: array
          items:
            $ref: "#/components/schemas/DailyCount"

    OverviewStats:
      type: object
      required: [totalCreators, totalScripts, totalViews, viewsToday]
      properties:
        totalCreators:
          type: integer
        totalScripts:
          type: integer
        totalViews:
          type: integer
        viewsToday:
          type: integer