# lib/api-zod/src/generated/types/accessLog.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FunScript Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface AccessLog {
  id: number;
  funScriptId: number;
  /** @nullable */
  ipAddress?: string | null;
  /** @nullable */
  userAgent?: string | null;
  /** @nullable */
  referrer?: string | null;
  accessedAt: string;
}