# lib/api-zod/src/generated/types/accessLogWithScript.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FunScript Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface AccessLogWithScript {
  id: number;
  funScriptId: number;
  /** @nullable */
  scriptTitle?: string | null;
  /** @nullable */
  shareToken?: string | null;
  /** @nullable */
  ipAddress?: string | null;
  /** @nullable */
  userAgent?: string | null;
  /** @nullable */
  referrer?: string | null;
  accessedAt: string;
}