# lib/api-zod/src/generated/types/creator.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FunScript Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface Creator {
  id: number;
  name: string;
  handle: string;
  /** @nullable */
  bio?: string | null;
  scriptCount?: number;
  totalViews?: number;
  createdAt: string;
  updatedAt?: string;
}