# lib/api-zod/src/generated/types/creatorInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FunScript Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface CreatorInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  handle: string;
  bio?: string;
}