# lib/api-zod/src/generated/types/creatorUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FunScript Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface CreatorUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  handle?: string;
  /** @nullable */
  bio?: string | null;
}