# lib/api-zod/src/generated/types/funScript.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FunScript Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface FunScript {
  id: number;
  creatorId: number;
  /** @nullable */
  creatorName?: string | null;
  /** @nullable */
  creatorHandle?: string | null;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  contentUrl?: string | null;
  shareToken: string;
  viewCount: number;
  /** @nullable */
  expiresAt?: string | null;
  createdAt: string;
  updatedAt?: string;
}