# lib/api-zod/src/generated/types/funScriptInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FunScript Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface FunScriptInput {
  creatorId: number;
  /** @minLength 1 */
  title: string;
  description?: string;
  contentUrl?: string;
  expiresAt?: string;
}