# lib/api-zod/src/generated/types/funScriptUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FunScript Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface FunScriptUpdate {
  /** @minLength 1 */
  title?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  contentUrl?: string | null;
  /** @nullable */
  expiresAt?: string | null;
}