# lib/api-zod/src/generated/types/overviewStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FunScript Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface OverviewStats {
  totalCreators: number;
  totalScripts: number;
  totalViews: number;
  viewsToday: number;
}