# lib/api-zod/src/generated/types/scriptStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FunScript Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { DailyCount } from './dailyCount';

export interface ScriptStats {
  totalViews: number;
  uniqueIps: number;
  dailyCounts: DailyCount[];
}